feeder/requirements.txt
pyuipc>=0.5
requests>=2.31.0
python-dotenv>=1.0.0
